LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-operations.https_rest-expected.txt


PASS promiseState helper works
PASS promiseStateFinal helper works
FAIL SLD(rollback) must detect InvalidStateError synchronously when chain is empty assert_equals: promise rejected on same task expected "rejected" but got "pending"
FAIL addIceCandidate must detect InvalidStateError synchronously when chain is empty assert_equals: expected "InvalidStateError" but got "Error"
PASS replaceTrack must detect InvalidStateError synchronously when chain is empty and transceiver is stopped
PASS createOffer uses operations chain
PASS Negotiationneeded only fires once operations chain is empty
FAIL setParameters does NOT use the operations chain assert_equals: Method is async expected "pending" but got "fulfilled"
PASS addIceCandidate chains onto SRD, fails before
PASS Operations queue not vulnerable to recursion by chained negotiationneeded
